pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "furseal-face-recognition"
version = "0.1.0"
description = "Fur seal face detection, recognition, and behavior analytics toolkit"
readme = "README.md"
authors = [
  { name = "Furseal Research Team" }
]
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.2.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "torchaudio>=2.2",
  "facenet-pytorch>=2.5.3",
  "opencv-python>=4.9",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "scipy>=1.11",
  "Pillow>=10.3",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "tqdm>=4.66",
  "tyro>=0.6.0",
  "pyyaml>=6.0",
  "rich>=13.7",
  "networkx>=3.2",
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "exifread>=3.0",
  "python-dotenv>=1.0",
  "faiss-cpu>=1.8.0",
  "plotly>=5.20",
  "pyarrow>=16.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "black>=24.4",
  "ruff>=0.5",
  "mypy>=1.10",
  "pre-commit>=3.7"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "C4"]
ignore = ["E203", "E266", "E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
